sandesh/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
pillow>=11.0.0
pandas==2.2.3
google-generativeai==0.3.2
beautifulsoup4==4.12.2
python-multipart==0.0.6